test: Move overflow tests to own crate
Need to optimize tests and disable overflow checks
to make the overflow tests perform acceptably.
